debian: import 1.10.0-1ubuntu2.1 from bionic-proposed to fix FTBFS
This change updates the package from 1.10.0-1ubuntu2 to
1.10.0-1ubuntu2.1 proper.

Since the change only fixes a build failure due to an expired key, the
updated upstream ubuntu package is only available in bionic-proposed,
and not in bionic-updates.

See https://bugs.launchpad.net/ubuntu/bionic/+source/gpgme1.0/+bug/1813581/comments/2
for an explanation about why it is kept only in -proposed.